Add a objectchooser to select images
We want use a ObjectChooser to select images from the journal
and specific directories, showing preview images.
Preview is not implemented in Sugar 0.96 in use in Haiti,
and we can't add specific directories to the standard ObjectChooser.
This patch copy the files needed from Sugar to the activity,
to have a local implememtation of the ObjectChooser with the needed functionality.