Implement a custom Chooser to select images
The ObjectChooser with preview do not run in Sugar 0.96
I have tried making copies of many tookit files in the activity directory
but was not possible.
A option is display a custom choser with preview
for the files in a particular directory,
and add a button to show the standard ObjectChooser.