Add categories filters for the image chooser
As showing all the images at the same time is very slow in the xo-1,
add the possibility to define a set of categories,
everyone with their own path. If the categories are defined,
a group of buttons is displayed, and after press the buttons,
the iconview is displayed showing the images in the selected directory.