Create epub files
A epub file is created with a page for the title and author,
and another big page with the content of the book.
We can separate the big page in individual pages for easier
navigation with the buttons in the xo.
The epub was validated with http://validator.idpf.org/