Allow use multiple paths by category and translate file names
We can show only a single directory using the datastore backend.
We also want be able to show and search by translated names
the solution is create a cached directory with symlinks to
the media found in the list of category_paths. If a dictionary data file
is found, the filenames for the links will be translated.