Fix filter in imagechooser when nothing is filtered
Previously we only created directories with links when translations
were available or multiple directories were configured for a category,
but that had a limitation, we can't search in the main directory
if there are not all the files in the same tree.
Then now when there are not translations or is a single directory,
we create  fake directory but only making a single symlink for
the directory, and not one for every file.
Other change needed is make a separated tree for backgrounds and
for actors, to not find images mixed in the searches.
That is done adding a image_type parameter to the ImageFileChooser.

All this have a limitation. As the links are created the first time
the user show a category, we will not get results in the search
until the images in the category are displayed at least one time.